Update for SMAPI 1.8, rename config fields, get rid of some code
