+++ AutomatedDoors/AutomatedDoors.cs
@@ -6,29 +6,22 @@ using StardewValley;
 using StardewValley.Buildings;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Input;
-using SMAPIAutomatedDoors;
 
 namespace AutomatedDoors
 {
 
     public class AutomatedDoors : Mod
     {
-        public static AutomatedDoorsConfig ModConfig { get; protected set; }
+        private AutomatedDoorsConfig _config;
 
         private bool openDoorsEventFired;
         private bool closeDoorsEventFired;
-        public int openDoorsTime;
-        public int closeDoorsTime;
-        public bool openRainyDays;
-        public bool openWinter;
-        public Dictionary<string, bool> buildings;
 
-        public override void Entry(params object[] objects)
+        public override void Entry(IModHelper helper)
         {
-            ModConfig = new AutomatedDoorsConfig().InitializeConfig(BaseConfigPath);
-            GameEvents.UpdateTick += Events_UpdateTick;
+            _config = Helper.ReadConfig<AutomatedDoorsConfig>();
             TimeEvents.DayOfMonthChanged += Events_NewDay;
-            Console.WriteLine("AutomatedDoors Loaded ...... [OK]");
+            GameEvents.OneSecondTick += Events_OneSecondTick;
         }
 
         public void Events_NewDay(object sender, EventArgs e)
@@ -37,16 +30,16 @@ namespace AutomatedDoors
             closeDoorsEventFired = false; 
         }
 
-        public void Events_UpdateTick(object sender, EventArgs e)
+        public void Events_OneSecondTick(object sender, EventArgs e)
         {
             if (!Game1.hasLoadedGame)
             {
                 return;
             }
             
-            if (!openDoorsEventFired && Game1.timeOfDay == ModConfig.openDoorsTime && Game1.IsWinter == ModConfig.openWinter)
-                {
-                    if (ModConfig.openRainyDays == true)
+            if (!openDoorsEventFired && Game1.timeOfDay == _config.timeDoorsOpen) //&& Game1.IsWinter == _config.openInWinter
+            {
+                    if (_config.openOnRainyDays == true)
                     {
                         OpenBuildingDoors();
                     }
@@ -55,7 +48,7 @@ namespace AutomatedDoors
                         OpenBuildingDoors();
                     }
                 }
-             if (!closeDoorsEventFired && Game1.timeOfDay >= ModConfig.closeDoorsTime)
+             if (!closeDoorsEventFired && Game1.timeOfDay >= _config.timeDoorsClose)
                 {
                     CloseBuildingDoors();
                 }
@@ -70,9 +63,9 @@ namespace AutomatedDoors
                 while (enumerator.MoveNext())
                 {
                     Building current = enumerator.Current;
-                    if ( !ModConfig.buildings.ContainsKey(current.nameOfIndoors) )
+                    if ( !_config.buildings.ContainsKey(current.nameOfIndoors) )
                     {
-                        ModConfig.buildings.Add(current.nameOfIndoors, true);
+                        _config.buildings.Add(current.nameOfIndoors, true);
 
                         if (current.animalDoorOpen == false)
                         {
@@ -80,9 +73,9 @@ namespace AutomatedDoors
                             openDoorsEventFired = true;
                         }
                     }
-                    else if (ModConfig.buildings.ContainsKey(current.nameOfIndoors))
+                    else if (_config.buildings.ContainsKey(current.nameOfIndoors))
                     {
-                        if (current.animalDoorOpen == false && ModConfig.buildings[current.nameOfIndoors] == true )
+                        if (current.animalDoorOpen == false && _config.buildings[current.nameOfIndoors] == true )
                         {
                             current.doAction(new Vector2(current.animalDoor.X + current.tileX, current.animalDoor.Y + current.tileY), Game1.player);
                             openDoorsEventFired = true;
@@ -90,9 +83,6 @@ namespace AutomatedDoors
                     }
                 }
             }
-
-            ModConfig.UpdateConfig<AutomatedDoorsConfig>();
-            ModConfig.WriteConfig();
         }
 
         public void CloseBuildingDoors()
